+++ ubelt/util_path.py
@@ -818,7 +818,6 @@ class Path(_PathBase):
             >>>     if 'CVS' in dirs:
             >>>         dirs.remove('CVS')  # don't visit CVS directories
         """
-        import os
         cls = self.__class__
         walker = os.walk(self, topdown=topdown, onerror=onerror,
                          followlinks=followlinks)
@@ -827,14 +826,18 @@ class Path(_PathBase):
 
     def __add__(self, other):
         """
-        Returns a new string that directly appends to the end of this fspath
-        representation.
+        Returns a new string starting with this fspath representation.
 
         Returns:
             str
 
         Allows ubelt.Path to be a better drop-in replacement when working with
         string-based paths.
+
+        Note:
+            It is not recommended to write new code that uses this behavior.
+            This exists to make it easier to transition existing str-based
+            paths to pathlib.
 
         Example:
             >>> import ubelt as ub
@@ -847,6 +850,32 @@ class Path(_PathBase):
         """
         return os.fspath(self) + other
 
+    def __radd__(self, other):
+        """
+        Returns a new string ending with this fspath representation.
+
+        Returns:
+            str
+
+        Allows ubelt.Path to be a better drop-in replacement when working with
+        string-based paths.
+
+        Note:
+            It is not recommended to write new code that uses this behavior.
+            This exists to make it easier to transition existing str-based
+            paths to pathlib.
+
+        Example:
+            >>> import ubelt as ub
+            >>> base = ub.Path('base')
+            >>> base_ = ub.Path('base/')
+            >>> base2 = ub.Path('base/2')
+            >>> assert 'foo' + base == 'foobase'
+            >>> assert 'foo' + base_ == 'foobase'
+            >>> assert 'foo' + base2 == str(base2.augment(dpath='foobase'))
+        """
+        return other + os.fspath(self)
+
     def endswith(self, suffix, *args):
         """
         Test if the fspath representation endswith a particular string
@@ -854,25 +883,65 @@ class Path(_PathBase):
         Allows ubelt.Path to be a better drop-in replacement when working with
         string-based paths.
 
+        Args:
+            suffix (str | Tuple[str, ...]):
+                One or more suffixes to test for
+
+            *args:
+                start (int): if specified begin testing at this position.
+                end (int): if specified stop testing at this position.
+
+        Returns:
+            bool: True if any of the suffixes are matched.
+
         Example:
             >>> import ubelt as ub
             >>> base = ub.Path('base')
             >>> assert base.endswith('se')
             >>> assert not base.endswith('be')
+            >>> # test start / stop cases
+            >>> assert ub.Path('aabbccdd').endswith('cdd', 5)
+            >>> assert not ub.Path('aabbccdd').endswith('cdd', 6)
+            >>> assert ub.Path('aabbccdd').endswith('cdd', 5, 10)
+            >>> assert not ub.Path('aabbccdd').endswith('cdd', 5, 7)
+            >>> # test tuple case
+            >>> assert ub.Path('aabbccdd').endswith(('foo', 'cdd'))
+            >>> assert ub.Path('foo').endswith(('foo', 'cdd'))
+            >>> assert not ub.Path('bar').endswith(('foo', 'cdd'))
         """
         return os.fspath(self).endswith(suffix, *args)
 
-    def startswith(self, suffix, *args):
+    def startswith(self, prefix, *args):
         """
         Test if the fspath representation startswith a particular string
 
         Allows ubelt.Path to be a better drop-in replacement when working with
         string-based paths.
 
+        Args:
+            prefix (str | Tuple[str, ...]):
+                One or more prefixes to test for
+
+            *args:
+                start (int): if specified begin testing at this position.
+                end (int): if specified stop testing at this position.
+
+        Returns:
+            bool: True if any of the prefixes are matched.
+
         Example:
             >>> import ubelt as ub
             >>> base = ub.Path('base')
             >>> assert base.startswith('base')
             >>> assert not base.startswith('all your')
+            >>> # test start / stop cases
+            >>> assert ub.Path('aabbccdd').startswith('aab', 0)
+            >>> assert ub.Path('aabbccdd').startswith('aab', 0, 5)
+            >>> assert not ub.Path('aabbccdd').startswith('aab', 1, 5)
+            >>> assert not ub.Path('aabbccdd').startswith('aab', 0, 2)
+            >>> # test tuple case
+            >>> assert ub.Path('aabbccdd').startswith(('foo', 'aab'))
+            >>> assert ub.Path('foo').startswith(('foo', 'aab'))
+            >>> assert not ub.Path('bar').startswith(('foo', 'aab'))
         """
-        return os.fspath(self).startswith(suffix, *args)
+        return os.fspath(self).startswith(prefix, *args)